refactor: restructure CityDB script execution and related utilities
